+++ source/mtree_int8_util.h
@@ -7,12 +7,12 @@
 
 #include "mtree_int8.h"
 
-int mtree_int8_distance_internal(mtree_int8* first, mtree_int8* second);
+long long mtree_int8_distance_internal(mtree_int8* first, mtree_int8* second);
 bool mtree_int8_equals(mtree_int8* first, mtree_int8* second);
 bool mtree_int8_overlap_distance(mtree_int8* first, mtree_int8* second, int distance);
 bool mtree_int8_contains_distance(mtree_int8* first, mtree_int8* second, int distance);
 bool mtree_int8_contained_distance(mtree_int8* first, mtree_int8* second, int distance);
 mtree_int8* mtree_int8_deep_copy(mtree_int8* source);
-int get_int8_distance(int size, mtree_int8* entries[size], int distances[size][size], int i, int j);
+long long get_int8_distance(int size, mtree_int8* entries[size], long long distances[size][size], int i, int j);
 
 #endif
